Check if views are correctly loaded before registering - Check if related commands are there - Ask user to reload VS Code window if not

+++ src/views/live-watch/live-watch.ts
@@ -17,6 +17,7 @@
 import * as vscode from 'vscode';
 import { DebugProtocol } from '@vscode/debugprotocol';
 import { GDBTargetDebugSession, GDBTargetDebugTracker } from '../../debug-session';
+import { vscodeViewExists } from '../../utils';
 
 export interface LiveWatchNode {
   id: number;
@@ -93,8 +94,11 @@ export class LiveWatchTreeDataProvider implements vscode.TreeDataProvider<LiveWa
         return this._activeSession;
     }
 
-    public activate(tracker: GDBTargetDebugTracker): void {
-        this.addVSCodeCommands();
+    public async activate(tracker: GDBTargetDebugTracker): Promise<boolean> {
+        if (!await this.addVSCodeCommands()) {
+            console.log('Live Watch: Live Watch window cannot be registered, abort activation');
+            return false;
+        }
         const onDidChangeActiveDebugSession = tracker.onDidChangeActiveDebugSession(async (session) => await this.handleOnDidChangeActiveDebugSession(session));
         const onWillStartSession =  tracker.onWillStartSession(async (session) => await this.handleOnWillStartSession(session));
         // Using this event because this is when the threadId is available for evaluations
@@ -116,6 +120,7 @@ export class LiveWatchTreeDataProvider implements vscode.TreeDataProvider<LiveWa
             onWillStartSession,
             onStackTrace,
             onWillStopSession);
+        return true;
     }
 
     public async deactivate(): Promise<void> {
@@ -135,7 +140,10 @@ export class LiveWatchTreeDataProvider implements vscode.TreeDataProvider<LiveWa
         });
     }
 
-    private addVSCodeCommands() {
+    private async addVSCodeCommands(): Promise<boolean> {
+        if (!await vscodeViewExists('liveWatch')) {
+            return false;
+        }
         const registerLiveWatchView = vscode.window.registerTreeDataProvider('cmsis-debugger.liveWatch', this);
         const addCommand = vscode.commands.registerCommand('vscode-cmsis-debugger.liveWatch.add', async () => await this.handleAddCommand());
         const deleteAllCommand = vscode.commands.registerCommand('vscode-cmsis-debugger.liveWatch.deleteAll', async () => await this.handleDeleteAllCommand());
@@ -170,6 +178,7 @@ export class LiveWatchTreeDataProvider implements vscode.TreeDataProvider<LiveWa
             addToLiveWatchFromVariablesViewCommand,
             showInMemoryInspectorCommand
         );
+        return true;
     }
 
     private async handleAddCommand() {
